Change app and library compose build plugins to share dependencies
