+++ build-plugin/src/main/kotlin/ProjectExtension.kt
@@ -1,6 +1,10 @@
 import org.gradle.accessors.dm.LibrariesForLibs
 import org.gradle.api.Project
+import org.gradle.kotlin.dsl.DependencyHandlerScope
 import org.gradle.kotlin.dsl.getByName
 
-val Project.libs: LibrariesForLibs
+internal val Project.libs: LibrariesForLibs
     get() = extensions.getByName<LibrariesForLibs>("libs")
+
+internal fun Project.dependencies(configuration: DependencyHandlerScope.() -> Unit) =
+    DependencyHandlerScope.of(dependencies).configuration()
