+++ build-plugin/src/main/kotlin/thunderbird.app.android.compose.gradle.kts
@@ -17,16 +17,7 @@ android {
 }
 
 dependencies {
-    val composeBom = platform(libs.androidx.compose.bom)
-    implementation(composeBom)
-    androidTestImplementation(composeBom)
+    configureSharedComposeDependencies(libs)
 
-    implementation(libs.bundles.shared.jvm.android.compose)
     implementation(libs.androidx.compose.activity)
-
-    debugImplementation(libs.bundles.shared.jvm.android.compose.debug)
-
-    testImplementation(libs.bundles.shared.jvm.test.compose)
-
-    androidTestImplementation(libs.bundles.shared.jvm.androidtest.compose)
 }
